partial class for gestures
